Split files and import them to other dart files
